Montgomery Emacs, from Joel N. Chiappa.
From https://minnie.tuhs.org/pipermail/tuhs/2021-April/023560.html

I have also put up the Montgomery Emacs for Unix:

  http://ana-3.lcs.mit.edu/~jnc/tech/unix/emacs

This is the version we were running on the 5th floor MIT V6 machine
(CSR), which by that point have absorbed a few V7isms (e.g. some
ioctl() stuff). So don't expect to be able to compile and run it,
without a fair amount of work. (I vaguely recall that it needs I+D
space, so maybe not on a /23 at all.) But at least the source is in C,
so you can read it. I don't think there's an un-modified version
online (i.e. the original Montgomery source), alas.